updating devfile for v2 schema
a new devfile was created to comply with v2 schema of DevSpaces (newer version of CodeREady Workspaces).
The liferay component previously an openshift type was changed to container type to accommodate those changes. So the container attribute was added with all the container info needed for liferay container to start from within the devspace to deploy the liferay helloworld app.

Also changes in all the commands were done to comply with new schema version.